update java to 21, make asyncRegistrarVariacao
