+++ infra/src/main/java/br/com/carteira/infra/integracoes/BMFBovespa.java
@@ -3,8 +3,4 @@ package br.com.carteira.infra.integracoes;
 public interface BMFBovespa {
 
     CotacaoDto getCotacao(String ticker);
-
-    String getErrorList();
-
-    String getLastErrorDetails();
 }
