+++ infra/src/main/java/br/com/carteira/infra/integracoes/CotacaoDto.java
@@ -1,6 +1,11 @@
 package br.com.carteira.infra.integracoes;
 
 public record CotacaoDto(
-        Double valor
+        Double valor,
+        boolean hasError,
+        String errorMessage
 ) {
+    public static CotacaoDto fromPrice(Double price) {
+        return new CotacaoDto(price, false, null);
+    }
 }
